Moved GPR parsing into own class. Fixed class cast exception in GPRParser. Moving to testing and fixing issues with BiGGIDs now
